+++ src/init.c
@@ -9,12 +9,14 @@ extern SEXP method_call_(SEXP, SEXP, SEXP);
 extern SEXP S7_class_(SEXP, SEXP);
 extern SEXP S7_object_(void);
 extern SEXP prop_(SEXP, SEXP);
+extern SEXP prop_set_(SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
     {"method_", (DL_FUNC) &method_, 4},
     {"method_call_", (DL_FUNC) &method_call_, 3},
     {"S7_object_", (DL_FUNC) &S7_object_, 0},
     {"prop_", (DL_FUNC) &prop_, 2},
+    {"prop_set_", (DL_FUNC) &prop_set_, 4},
     {NULL, NULL, 0}
 };
 
@@ -29,6 +31,9 @@ SEXP sym_abstract;
 SEXP sym_constructor;
 SEXP sym_validator;
 SEXP sym_getter;
+
+SEXP sym_dot_should_validate;
+SEXP sym_dot_setting_prop;
 
 SEXP ns_S7;
 
@@ -48,11 +53,8 @@ void R_init_S7(DllInfo *dll)
     sym_constructor = Rf_install("constructor");
     sym_validator = Rf_install("validator");
     sym_getter = Rf_install("getter");
+    sym_dot_should_validate = Rf_install(".should_validate");
+    sym_dot_setting_prop = Rf_install(".setting_prop");
 
     ns_S7 = Rf_findVarInFrame(R_NamespaceRegistry, Rf_install("S7"));
-
 }
-
-
-
-